LCD screen bug fix and optimization
Fixed a bug in the function screen_write_text_aligned() that caused flickering of the top-left screen characters when there was text on the same screen aligned to the right side. Optimized text alignment to the right side. Now, the end pixel for this type of alignment is 128, instead of 126 as it was before this update.